Introduce Connector and availability for Connection
Problem
The DuplexConnection class doesn't expose any way to probe its state,
this is somewhat problematic and doesn't fit well in the current
model where the availability is the composable way of chaining state.

The ReactiveSocketFactory create a ReactiveSocket from an `address`
(currently SocketAddress). There are places in the code where the concept
of address is not needed and it leads to implicit dependency on it.

Solution
Add a `double availability()` method to the DuplexConnection interface,
most implementation of this method will be straightforward (just
returning 0.0 if the underlying resource is unavailable, 1.0 otherwise).

Split the concept of ReactiveSocketFactory in two, the Factory and the
Connector. The Connector is responsible for creating the ReactiveSocket
from the address, and the factory can create a ReactiveSocket without
any argument.