Propagate cancellation when a TransportException occurs
Problem
When a TransportException happens on the transport, it is propagated up
in the chain but the Susbscription is not cancelled. Then, any source
associated with the chain continue to emit events.

One example of that is the Keep-Alive Observable, which regularly emit
keep-alive messages at fixed interval, when the connection is closed, the
observable has to be cancelled.

Solution
In all DuplexConnection implementation, cancel the Subscription when we
saw a TransportException.