Refactor Factory to Connector + Implement availability (#15)
* Refactor Factory to Connector + Implement availability

Problem
We need to comply with the recent refactoring "Factory to Connector" introduced
in reactivesocket-java. Also, implementation of DuplexConnection have to return
an availability.

Solution
The Factory to Connector refactoring is pretty straight forward.

All implementations of DuplexConnection return an availability which directly
map the state of the underlying resource (0.0 if the resource is closed, 1.0
otherwise). This will greatly help the load-balancer to select a valid
connection.

I removed the blocking method from reactivesocket-java, so I moved some utility
blocking code inside the TestUtil class.

Clean-up of the Netty implementation, remove unused args, explicitey specify
tcp socket configuration.

Shortened the toString name to ease log-reading.

Bug
All DuplexConnection implementations now cancel the subscription when an exception
occurs. This fix the problem, where the Keep-Alive Observable kept trying to
send a keep-alive on a closed connection.

* Refactor the TestUtil helper to return CompletableFuture

* rs version

* Restore rs dependency to latest.release

* Remove mavenLocal in gradle build file

* Propagate cancellation in all cases

* Use helper method from ReactiveSocket Unsafe