Migrate reactivesocket-load-balancer into reactivesocket-java
**Problem**
The multi-repo structure is difficult to use in a context of rapid
short iterations.

**Solution**
Migrate the reactivesocket-load-balancer repo inside reactivesocket-java.
The submodules have been renamed like this:

- reactivesocket-load-balancer-builder -> reactivesocket-client
- reactivesocket-load-balancer-core -> reactivesocket-client
- reactivesocket-load-balancer-eureka -> reactivesocket-discovery-eureka
- reactivesocket-load-balancer-servo -> reactivesocket-stats-servo

I also added a simple module `reactivesocket-examples`, which only contains
one file so far. The goal of this module is to demonstrate simple examples.