Unsubscribe from the connection before `shutdown` the RS.
**Problem**
When we close a ReactiveSocket, it closes the underlying DuplexConnection
which trigger an unnecessary onError on the Requester/Responder.

**Solution**
Store the disposable of the DuplexConnection subscription as an instance
variable, and dispose it before closing the DuplexConnection.

**Modification**
I also added a `name` method in Requester/Responder to simplify debugging
code where we have client and server in the same JVM.
It now shows which Requester/Responder is generating an error, the client's
one or the server's one.