Completing subscriber after sending error frame (caused memory leak).
#### Problem

In some cases, `Responder` was converting an `onError()` to `onNext(Frame.Error)` but was not sending an `onComplete()` after the `onNext()`.
This causes the `subscriber` to wait for a terminal event, which never arrives, or if it arrives, does not find the `streamId` to be valid since the code was invoking `cleanUp()`.
This although subtle, but causes memory build up as the `Subscriber` in the case of a server, is the subscriber which is writing the response. If the response does not complete, all state associated with the request is kept alive forever.

#### Solution

Call `onComplete()` after sending an error frame.

This code does not call `cancel()` after `onComplete()` as reactive streams spec does not mandate cancellation post a terminal event, unlike `RxJava`.

#### Result

No more memory leak!