Loadbalancer: Refactor socket/factory Map into Lists + bugfix. (#112)
* Loadbalancer: Refactor socket/factory Map into Queues + bugfix.

**Problem**
The loadbalancer uses internally 2 maps, one for the list of `ReactiveSocketFactory`
it can use for creating a new `ReactiveSocket`, and one for the active `ReactiveSocket`.
The link is made between the 2 maps through the `SocketAddress` of the associated
remote server. This is kind of clunky and requires explicit knowledge of
the `SocketAddress` (result type of `remote()`).

Also, when we select a new factory to connect to or when we select the slowest
ReactiveSocket to quick out, the sorting is made via the `sorted()` method
on the java stream. The comparison function is actually unstable because a
tcp socket can be closed while we do the sorting.

**Solution**
Use two queues, one for factories and one for ReactiveSocket. Selecting a factory
is made using the "Power of 2 Choices" technique, and we also "age" the unselected
factories by moving them at the end of the queue. The WeightedSocket now contains
a reference to the associated factory, and closing it add back the factory to the
factory queue.

**Modification**
Move the `WeightedSocket` class inside the `LoadBalancer`, no

I created a proper constructor to `LoadBalancer` with javadoc explaining all
parameters the algorithm requires.

The `LatencySubscriber` inside the LoadBalancer treats some exceptions
particularly:

- `TimeoutException` latency is used for predicating the next latency
- `TransportException` & `ClosedChannelException` removes the ReactiveSocket
   from the active list.

`Publishers.onError` now also `cancel` the subscription.

I created a StressTest file, which stress most of the part of the Client:

- Adding/Removing new servers
- High request concurrency
- Dealing with failing/black-hole servers