Loadbalancer: Move inter-arrival time as constant.
***Problem***
Hard-coded values in the code.

***Solution***
Move those hard-coded values into static members.

Initial inter-arrival time is 1 seconds (The average time between
two requests for a particular server). This value will be updated continuously.

Inter-arrival factor is now 500, after 500 times the default arrival time
we'll artificially decrease the load of a server so that it is selected
again. It solves the problem of server becoming temporarly latent, and
never being selected again.