Loadbalancer: Only reduce aperture based on active sockets (not pending).
***Problem***
The decision of reducing the aperture is based on the number of active sockets
+ the number of pending socket. This could be bad as we close sockets before
the pending ones connect, which should leave us under the aperture target.

***Solution***
Only reduce the aperture when the number of *active* socket is below the
aperture.

***Modification***
I also increase the aperture only when we have available factories in
order to reduce the debug log.
Reduce the log level of these logs to DEBUG.