Generics correctness in `ClientBuilder`
#### Problem

`ClientBuilder.withSource()` takes an argument of `Publisher<Collection<T>>` which means it doesn't accept an `Publisher<List<T>>`.
`Eureka` is providing `Publisher<List<T>>` so it is incompatible with `ClientBuilder`

#### Modification

Changed `ClientBuilder.withSource()` to take `Publisher<? extends Collection<T>>`.
Also changed `Eureka` to return `Publisher<Collection<T>>` which is the correct thing to do as returning a `List` doesn't really make a difference.