Publishers cleanup and new functions
#### Problem

`Publishers.retry()` had the following bugs:

- Retry subscription was sending duplicate `onSubscribe` callbacks to the downstream subscribers. This is illegal according to spec.
- Flow control was not properly managed if the error happens after a few emissions.

Since we have removed retry functionality from the `ClientBuilder` there is no reason to maintain this code.

`Publishers` was not wiring the downstream subscriber cancellation to the cancellation of the `Subscriber`. This means that a cancel followed by `onNext` would not have stopped the emission, which although not necessary (because of the inherent race between cancel and emissions) but is good to have.

 #### Modification

 - Removed `RetrySocket` and `Publishers.retry()` since they are not used.
 - Added (moved from `PublisherUtils`) `.empty()`, `.just()` and `.error()` to `Publishers`
 - Added `concatEmpty()` to `Publishers`. This is required in places where we concat `Publisher<Void>` in custom crafter code. (I will send another PR which is to use this operator in `ReactiveSocket.close()`
 - Added tests for functions in `Publishers`.