Subscriber cleanup
#### Problem

`Publishers` was using different implementations of `CancellableSubscriber` to provide different callbacks using inheritance. eg: by providing methods like `doOnNext`, `doOnCancel`, etc.

This is problematic as it requires calls to `super.do*` by implementator. Also, it created unnecessary class hierarchy with little benefit.

#### Modification

Rolled up all implementations into a single class that takes various callbacks as different lambdas.
This is easier to use and there are less chances of errors when the implementor forgot to call `super` methods.

Thanks @yschimke for the comment in PR #127 which led me to change the approach.