Fix Error flags to match the spec + review the `frame` package.
***Problem***
There is an error in the Error frame, the error codes used for `APPLICATION_ERROR`,
`REJECTED`, `CANCELED` and `INVALID` were invalid.

***Solution***
Update the constants.
I don't think it requires doing any version bump, since the error code was not used
prior to #150.

***Modification***
I also took the oportunity to review the `frame` package and update it to the Java
code style.