Fix Responder leak in LeaseGovernor
***Problem***
There's currently a leak of responder reference in the `LeaseGovernor`,
upon disconnection, the `Responder` is not removing its reference from the
Governor.

***Solution***
On cancel/disconnection/exception the Responder is now removing itself.

I tested that with the JS and Java client.
The `TestConnection` doesn't make disconnection testing possible with the
present code, I'll improve this in another commit.

***Modification***
I also clean-up the code-style of PublisherUtils.