More fairness for the FairLeaseGovernor (#157)
***Problem***
The `FairLeaseGovernor` currently distributes its budget among the connected
peers. It does so by dividing its total budget by the number of connected clients,
and add the remaining budget (`n`) among the first `n` responders.

This is fine when the budget is far greater than the number of clients, but in the
opposite situation, it leads to client starving, as the budget is always distributed
in the same order.

***Solution***
Shuffle the list of client prior to distributing the budget.