Extract interface `AvailabilityProvider`
#### Problem

`DuplexConnection`, `ReactiveSocket` and `ReactiveSocketClient` all provide a method `double availability()` and so can benefit from having a common interface.
This will be useful for giving event callbacks that retrieve availability as availability follows a pull model instead of push.

#### Modification

Added a new interface `AvailabilityProvider` and have the other interfaces extend it.

#### Result

Common way to access availabilty.