Respond to keep-alive with response flag off.
#### Problem

`ServerReactiveSocket` upon receiving a `KeepAlive` frame responds with a `KeepAlive` frame as an ack. This ack MUST not have respond flag set to true. Doing so will result in an infinite loop of keep-alive frames sent between the peers.

#### Modification

`ServerReactiveSocket` responds with the respond flag turned off.

#### Result

No infinite loop of keep-alive frames.