added sliding window histogram (#200)
Added sliding window histogram

Problem
Current histogram doesn't resets on every get on every dimension. This creates choppy metrics.

Modifications
Modified the timer to have n histrograms in a queue and rotate of over them to smooth out the histogram over a sliding window. Switched the dimensions of the histrogram to a label called value. Switched the reseting of the histrogram to a closure that is called by the timer instead so it will rotate the histrogram once per window.

Result
Correct metrics over a rolling window.