Empty stacktrace of `TimeoutException` (#202)
Empty stacktrace of `TimeoutException`

#### Problem

`TimeoutPublisher` creates a new `TimeoutException` instance for every timeout. This is costly when the system is under load and has a lot of timers that timeout, which is typical when handling latent outbound services.

#### Modification

Used a cached `TimeoutException` instance and empty the stacktrace of the exception.

#### Result

Less overhead of timeouts.