Using native spectator counter and timer
__Problem__

Currently we have custom implementation of timers and counters.
These implementations are somewhat not recognized inside Netflix and hence does not get published to our metrics system.
At this point it is better to use the native counterparts than to debug the issue, we can revert to the custom implementation when we see performance issues.

 __Modification__

 Replace custom usages with native spectator classes.

 __Result__

 Metrics correctly published.