Make Aeron compile again (#310)
On way to working aeron again.  Compiling but still failing.

"In order to use Aeron someone needs to fix the handshaking in the beginning. Basically what needs to happen is the acking that the connection is established needs to happen in the AvailableImageHandler. After the first connection subsequent connections will fail right now."