Improvements to Fragmentation and Reassembly
This change improves the implementation and test coverage of the fragmentation
and reassembly functionality.  The fragmentation functionality now support all
frame types (previously, only the PAYLOAD frame could be fragmented) and
ensures that a fragmentation and reassembly roundtrip results in an identical
frame.

The implementations of the fragmentation and reassembly functionality has been
implemented on top of the new framing abstraction for testability and to start
moving off of the abstraction that leaks frame lengths and stream ids.  Note
that the implementation of DuplexConnection *does not* move to the new framing
abstraction.  As this interface is not yet migrated, the
FragmentationDuplexConnection is responsible for translating back and forth
between the old abstraction and the new one.  This has an implication that
overhead will increase (although minimally because of zero-copy data and the
use of Flyweights) temporarily during the migration to the new abstraction.
Once that abstraction has been adopted everywhere, the translation will be
removed, and the overheard will go back to it's current levels.