Fragmentation (#481)
* Alternate Framing Abstraction

This change introduces a framing package that contains strongly typed
representations of all of the RSocket frame types.  These representations are
introduced as a separate abstraction because they specifically remove the
frame length and stream id that are an optional, transport-layer concern. This
allows will allow the main body of the code to be gradually moved over to an
implementation that only deals with the logical frame types, and pushes the
decoration of the frames with frame length and/or stream id down to the
transport layer.  Note that no implementation has been converted over; this is
just the prework to have an the frame types ready for migration.

These frame type representations are all implemented using the Flyweight
pattern (i.e. use the Netty Recycler) and preserve the zero-copy nature of the
existing frame representation.  These implementations have consistent and
symmetric management of ByteBuf retain and release lifecycle, in addition to
well documented and proper nullability enforced APIs, specific to each frame
type. Each implementation is also tested against raw binary representations to
ensure the strictest specification compliance.

Finally, improvements were made to testing with regards to configurable test
logging, an AssertJ ByteBuf Representation for visual comparison on test
failure, and other conveniences.

* Improvements to Fragmentation and Reassembly

This change improves the implementation and test coverage of the fragmentation
and reassembly functionality.  The fragmentation functionality now support all
frame types (previously, only the PAYLOAD frame could be fragmented) and
ensures that a fragmentation and reassembly roundtrip results in an identical
frame.

The implementations of the fragmentation and reassembly functionality has been
implemented on top of the new framing abstraction for testability and to start
moving off of the abstraction that leaks frame lengths and stream ids.  Note
that the implementation of DuplexConnection *does not* move to the new framing
abstraction.  As this interface is not yet migrated, the
FragmentationDuplexConnection is responsible for translating back and forth
between the old abstraction and the new one.  This has an implication that
overhead will increase (although minimally because of zero-copy data and the
use of Flyweights) temporarily during the migration to the new abstraction.
Once that abstraction has been adopted everywhere, the translation will be
removed, and the overheard will go back to it's current levels.