Migration of Spectator to Micrometer (#486)
This change migrates the metrics-gathering project from using Spectator to
using Micrometer (https://micrometer.io).  This change fully moves both the
DuplexConnectionInterceptor and RSocketInterceptor to a Micrometer-based
implementation, changing some of the metric names.  The metric name changes
were made at the advice of the project lead of Micrometer in order to get the
most useful (and compatible) metrics and dimensions for monitoring an RSocket-
based system.

As part of writing the tests, some necessary changes were detected in the
framing abstraction.  These changes are included in this.

[resolves #434]