Fix Reference Count Leak
Previously, there were a couple of frames that converted Strings to ByteBufs
automatically as part of construction.  Those frames though, retained any
ByteBufs passed into them so that they could be released by the creators as
necessary.  Since the automatic construction was not doing this release on the
String-based ByteBufs, there was a minor leak of references.  This change
updates those locations to release the String-based ByteBufs and accurately
account for their retention and release.