Bintray Publish Version
Previously, an update to the build system forgot to include a flag to the
Bintray publication of tags that used the tag value as the version being
published.  This change updates that Travis script to fix this oversight.

In addition, a minor change was made to the Bintray configuration to match the
tagging scheme used by the project.