Extract Project Version
Previously the project version in the build was explicitly specified within
the build script itself.  This way of specifying the project version makes the
value non-overrideable from the command line.  This change moves that value to
the gradle.properties file where it becomes a "normal" property, overridable
using the standard -P syntax.