Upload to Bintray and Maven Central
Previously, an update to the build system took a stab at uploading releases to
Bintray and them promoting them to Maven Central.  Unsurprisingly, without a
release to test against, this didn't work on the first try.  The first change
required a move of the bintray configuration from the top-level project to
each subproject that would be published.  Unlike before, this is now a subset
of projects excluding things like the integration tests.  Second, this change
finishes improvements to the Maven POM in order to meet the Maven Central
requirements for publication.  Specifically three developers were added to the
POMs in order for Sonatype to reach them if necessary.