Slow Test Run Configuration (#512)
Currently the TransportTest#requestChannel2_000_000 test (for all transports)
takes a very long time on some systems.  It's a useful test to have, but not
necessarily to run at all times.  This change creates an @SlowTest annotation
that is meta-annotated so that tests will only be enabled if
TEST_SLOW_ENABLED=true has been set.

[resolves #511]