Avoid moving full buffer's readerIndex when slicing entries
This implies that the index from which to decode and slice the next
entry MUST be provided by the user. To that end, we provide a method
to compute the next entry's index, given the pair of ByteBuf slices
returned by decodeMimeAndContentBuffers.

Also renamed that method to decodeMimeAndContentBufferSlices, to make
it more obvious that it produces slices (readerIndex impact and slices
nature are mentioned in the javadoc).

CompositeMetadata now keeps the nextEntryIndex as state and uses that
for hasNext / decodeNext.

Signed-off-by: Simon Baslé <sbasle@pivotal.io>