CompositeMetadata as Iterable
Previously, the implementation of CompositeMetadata resulted in a
mutable type that didn't fit in with either imperative (for) or reactive
(fromIterable) iteration strategies.  This change updates the type to
implement Iterable, and return an Iterator that lazily traverses the
ByteBuf.

There are also some improvements to the Flyweight in order to support
this iteration style.

Signed-off-by: Ben Hale <bhale@pivotal.io>