Well known mime types and Composite metadata extension (#635)
* Add encodeUnsignedMedium to NumberUtils

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Add WellKnownMimeType enum with String/ID conversions

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Add a flyweight capable of encoding Composite Metadata

The flyweight also exposes a method to decode Composite Metadata to a
Map.

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Move flyweight into metadata package, change decodeToMap to decodeNext

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Add CompositeMetadata interface with factory methods to decode/encode

The API is intended as a readonly way of listing the composite metadata
mime-buffer pairs, as well as accessing metadata pairs by mime type or
index, including the case where a mime type is associated with several
metadata entries (`List<Entry> getAll(String)`)...

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Add first test for CompositeMetadata (decode smoke test)

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Expose incremental decode method that decodes a single metadata entry

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* WellKnownMimeType parsing now avoids exceptions (except null arguments)

Instead has 2 special enums. int parsing allows detection of potentially
legit type that is still unknown to the current implementation, which
could help making it more future-proof.

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* rework encoding / decoding API around Entry, accommodate 3 mime case

Encoding becomes a mirror of decoding, purely based on Entry.

Accommodate 3 flavors of mime type decoding:

 - compressed well known with id matching an enum
 - uncompressed custom string
 - compressed but known as reserved for future use (future proofness)

 The later can be decoded into a special Entry that will allow re
 encoding it and transmitting it to other nodes as it was.

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* No need for a CompositeMetadata interface, single concrete implem

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Expose the CompositeMetadataFlyweight for low-level, simplify API

 - the flyweight deals with low-level encoding/decoding with
 minimal garbage and only uses exceptions in the encoding path
 -- encoding a non-ASCII custom mime type OR
 -- encoding an out of range byte
 - the CompositeMetadata is an _option_ for a higher level
 Iterable-based API, but which instantiates new types

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Remove CompositeMetadata, make Entry child of flyweight

Entry is now the only high-level abstraction, will need a way to
decode into a List<Entry>.

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Add javadoc for CompositeMetadataFlyweight methods

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Switch to a single Entry implementation, remove isPassthrough

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Rename Entry#decodeEntry to #decode, add #decodeAll

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* apply Google-style formatting

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Use precomputed array for WellKnownMimeType.fromId

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Use precomputed Map<String,WellKnownMimeType> for fromMimeType + jmh benchmark to validate perf increase

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Several custom mime metadata header encoding improvements:

 - avoid the use of a CompositeByteBuf
 - length can be safely predicted due to requirement that custom mime
 type be ASCII only
 - still using UTF8 writing and ByteBufUtil.isText(ASCII) to detect
 non-ascii text as:
   - writeAscii would not reject non-ascii chars but replace with `?`
   - ISO-8859 chars (eg. Latin-1) could be considered valid when only
   checking the number of written bytes
  - use of ByteBufUtil.writeUtf8 to simplify encoding the mime into the
  preallocated buffer
 - release the buffer if there is an encoding exception

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Use Unpooled buffers in tests

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Remove Entry, replace with CompositeMetadata Iterator-like abstraction

The CompositeMetadata wraps a ByteBuf and exposes a hasNext/decodeNext
API (reminiscent of the Iterator API) as a facade over the flyweight's
methods.

Each round decodeNext() is correctly invoked, the CompositeMetadata
state is updated and the getters can be used to retrieve decoded entry
components.

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Add an encode flyweight method that attempts compressing String

This is a qol method that will first check if the String passed can be
mapped to a WellKnownMimeType (and thus compressed to a 1 byte
representation).

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* Avoid moving full buffer's readerIndex when slicing entries

This implies that the index from which to decode and slice the next
entry MUST be provided by the user. To that end, we provide a method
to compute the next entry's index, given the pair of ByteBuf slices
returned by decodeMimeAndContentBuffers.

Also renamed that method to decodeMimeAndContentBufferSlices, to make
it more obvious that it produces slices (readerIndex impact and slices
nature are mentioned in the javadoc).

CompositeMetadata now keeps the nextEntryIndex as state and uses that
for hasNext / decodeNext.

Signed-off-by: Simon Baslé <sbasle@pivotal.io>

* CompositeMetadata as Iterable

Previously, the implementation of CompositeMetadata resulted in a
mutable type that didn't fit in with either imperative (for) or reactive
(fromIterable) iteration strategies.  This change updates the type to
implement Iterable, and return an Iterator that lazily traverses the
ByteBuf.

There are also some improvements to the Flyweight in order to support
this iteration style.

Signed-off-by: Ben Hale <bhale@pivotal.io>

* Apply googleFormat

Signed-off-by: Simon Baslé <sbasle@pivotal.io>